refactor(mattermost): own empty-body policy in helpers
